requirements.txt
django==1.11
djangorestframework==3.6.2
psycopg2==2.7.1
channels==1.1.3